+++ morpheus-graphql-core/src/Data/Morpheus/Types/App.hs
@@ -55,16 +55,16 @@ import Data.Morpheus.Types.Internal.Resolving
   ( ResolverContext (..),
     ResponseStream,
     ResultT (..),
-    RootResModel,
+    RootResolverValue,
     cleanEvents,
     resultOr,
-    runRootResModel,
+    runRootResolverValue,
   )
 import Data.Morpheus.Types.Internal.Stitching (Stitching (..))
 import Data.Morpheus.Validation.Document.Validation (ValidateSchema (..))
 import Relude hiding (empty)
 
-mkApp :: ValidateSchema s => Schema s -> RootResModel e m -> App e m
+mkApp :: ValidateSchema s => Schema s -> RootResolverValue e m -> App e m
 mkApp appSchema appResolvers =
   resultOr
     FailApp
@@ -87,7 +87,7 @@ instance Monad m => Semigroup (App e m) where
 
 data AppData event (m :: * -> *) s = AppData
   { appConfig :: Config,
-    appResolvers :: RootResModel event m,
+    appResolvers :: RootResolverValue event m,
     appSchema :: Schema s
   }
 
@@ -108,7 +108,7 @@ runAppData ::
 runAppData AppData {appConfig, appSchema, appResolvers} request = do
   validRequest <- validateReq appSchema appConfig request
   resovers <- withSystemFields (schema validRequest) appResolvers
-  runRootResModel resovers validRequest
+  runRootResolverValue resovers validRequest
 
 validateReq ::
   ( Monad m,