+++ morpheus-graphql-core/src/Data/Morpheus/Types/Internal/Resolving.hs
@@ -4,7 +4,7 @@
 module Data.Morpheus.Types.Internal.Resolving
   ( Resolver,
     LiftOperation,
-    runRootResModel,
+    runRootResolverValue,
     lift,
     Eventless,
     Failure (..),
@@ -14,14 +14,14 @@ module Data.Morpheus.Types.Internal.Resolving
     Result (..),
     ResultT (..),
     unpackEvents,
-    ObjectResModel (..),
-    ResModel (..),
+    ResolverObject (..),
+    ResolverValue (..),
     WithOperation,
     PushEvents (..),
     subscribe,
     ResolverContext (..),
     unsafeInternalContext,
-    RootResModel (..),
+    RootResolverValue (..),
     resultOr,
     withArguments,
     -- Dynamic Resolver
@@ -39,7 +39,7 @@ module Data.Morpheus.Types.Internal.Resolving
     ResolverState,
     liftResolverState,
     mkValue,
-    FieldResModel,
+    ResolverEntry,
     sortErrors,
     EventHandler (..),
   )
@@ -61,27 +61,29 @@ import Data.Morpheus.Types.Internal.Resolving.Core
 import Data.Morpheus.Types.Internal.Resolving.Event
 import Data.Morpheus.Types.Internal.Resolving.Resolver
 import Data.Morpheus.Types.Internal.Resolving.ResolverState
+import Data.Morpheus.Types.Internal.Resolving.ResolverValue
+import Data.Morpheus.Types.Internal.Resolving.RootResolverValue
 import qualified Data.Vector as V
   ( toList,
   )
 import Relude
 
-mkString :: Token -> ResModel o e m
+mkString :: Token -> ResolverValue m
 mkString = ResScalar . String
 
-mkFloat :: Double -> ResModel o e m
+mkFloat :: Double -> ResolverValue m
 mkFloat = ResScalar . Float
 
-mkInt :: Int -> ResModel o e m
+mkInt :: Int -> ResolverValue m
 mkInt = ResScalar . Int
 
-mkBoolean :: Bool -> ResModel o e m
+mkBoolean :: Bool -> ResolverValue m
 mkBoolean = ResScalar . Boolean
 
-mkList :: [ResModel o e m] -> ResModel o e m
+mkList :: [ResolverValue m] -> ResolverValue m
 mkList = ResList
 
-mkNull :: ResModel o e m
+mkNull :: ResolverValue m
 mkNull = ResNull
 
 unPackName :: A.Value -> TypeName
@@ -89,9 +91,9 @@ unPackName (A.String x) = TypeName x
 unPackName _ = "__JSON__"
 
 mkValue ::
-  (LiftOperation o, Monad m) =>
+  (Monad m) =>
   A.Value ->
-  ResModel o e m
+  ResolverValue m
 mkValue (A.Object v) =
   mkObject
     (maybe "__JSON__" unPackName $ HM.lookup "__typename" v)