+++ src/Data/Morpheus/Server/Types/GQLType.hs
@@ -50,8 +50,7 @@ import Data.Morpheus.Server.Types.SchemaT
     TypeFingerprint (..),
   )
 import Data.Morpheus.Server.Types.Types
-  ( MapKind,
-    Pair,
+  ( Pair,
     Undefined (..),
   )
 import Data.Morpheus.Types.ID (ID)
@@ -271,8 +270,9 @@ instance GQLType a => GQLType (Set a) where
   type KIND (Set a) = WRAPPER
   __type _ = __type $ Proxy @[a]
 
-instance (Typeable k, Typeable v) => GQLType (Map k v) where
+instance (GQLType k, GQLType v, Typeable k, Typeable v) => GQLType (Map k v) where
   type KIND (Map k v) = WRAPPER
+  __type _ = __type $ Proxy @[Pair k v]
 
 instance GQLType a => GQLType (Resolver o e m a) where
   type KIND (Resolver o e m a) = WRAPPER
@@ -287,6 +287,3 @@ instance GQLType b => GQLType (a -> b) where
   __type _ = __type $ Proxy @b
 
 instance (Typeable a, Typeable b, GQLType a, GQLType b) => GQLType (Pair a b)
-
-instance (Typeable a, Typeable b, GQLType a, GQLType b) => GQLType (MapKind a b m) where
-  __type _ = __type $ Proxy @(Map a b)