+++ morpheus-graphql-core/changelog.md
@@ -8,6 +8,8 @@
 
 ### Breaking changes
 
+- `RootResModel` was renamed to `RootResolverValue`
+- `ResModel` was replaced with more general `ResolverValue`
 - `GQLScalar` was replaced with `EncodeScalar` and `DecodeScalar` type-classes.
 
 - `Value.Float` is now `Double` instead of `Float`.